requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pycodestyle==2.12.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
